Run kubernetes deployment as non-root
We don't need to specify `runAsUser` because the USER in the Dockerfile
is the numeric user 1000 and kubernetes will be able to assert that the
USER isn't root (0).

When you use a non-numeric USER in the Dockerfile, you need to specify
`runAsUser` because kubernetes can't make sure that the specified user
isn't root just by inspecting the docker image.

But using `runAsUser` isn't desirable because it complicates things, in
our case we have a mix of kops(in aws) and openshift clusters and for
openshift it injects the `runAsUser` automatically and if it's not in
the random range assigned to the namespace, it fails, and the random
range for user ids is usually very high (> 1000000000).